// test/integration/test/ports/prisma/functional/json-fields/_fixture/generated/contract.d.ts
export type StorageHash =
  StorageHashBase<'c70bd8930c375d443f8971c1fc6e78261fde3d64bfec53060850fa5f9732991d'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Entry: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly json: CodecTypes['pg/json@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Entry: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly json: CodecTypes['pg/json@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly entry: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly json: CodecTypes['pg/json@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly entry: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly json: CodecTypes['pg/json@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly entry: {
              columns: {
                readonly id: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly json: {
                  readonly nativeType: 'json';
                  readonly codecId: 'pg/json@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly entry: { readonly namespace: 'public' & NamespaceId; readonly model: 'Entry' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Entry: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly json: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/json@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'entry';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly json: { readonly column: 'json' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;